Pesquisa list all done

+++ IFPRAcessoMain/models.py
@@ -9,7 +9,7 @@ class Pessoa(models.Model):
     nome_pessoa = models.CharField(max_length=50, unique=True)
     id_pessoa = models.ForeignKey(Identificador, on_delete=models.PROTECT)
     cracha_pessoa = models.IntegerField(unique=True)
-    matricula_pessoa = models.IntegerField(unique=True)
+    matricula_pessoa = models.BigIntegerField(unique=True)
     ano_entrada = models.IntegerField()
     ativo = models.CharField(max_length=1, default="S")
     def __str__(self):
